fixed variable issue

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-	int i, j, k, l;
+	long int i, j, k, l;
 
 	j = 1;
